% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_behav_surf}
\alias{tbl_behav_surf}
\title{Surface Behavior Data (processed)}
\format{An object of class \code{"data.frame"};}
\usage{
data(tbl_behav_surf)
}
\description{
Processed surfaced behavior data from satellite telemetry deployments on
harbor seals in the Aleutian Islands, Alaska from 2014-2017
}
\keyword{datasets}
