% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_locs}
\alias{tbl_locs}
\title{Location Data (processed)}
\format{An object of class \code{"data.frame"};}
\usage{
data(tbl_locs)
}
\description{
Processed location data from the satellite telemetry deployments on
harbor seals in the Aleutian Islands, Alaska from 2014-2017
}
\keyword{datasets}
