% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_percent}
\alias{tbl_percent}
\title{Percent Timeline Data (processed)}
\format{An object of class \code{"data.frame"};}
\usage{
data(tbl_percent)
}
\description{
Processed percent timeline data from satellite telemetry deployments on
harbor seals in the Aleutian Islands, Alaska from 2014-2017
}
\keyword{datasets}
